+++ backend/canisters/user_index/impl/src/updates/upgrade_user_canister_wasm.rs
@@ -18,7 +18,10 @@ async fn upgrade_user_canister_wasm(args: Args) -> Response {
         local_user_index_canisters,
     } = match read_state(|state| prepare(args, state)) {
         Ok(ok) => ok,
-        Err(response) => return response,
+        Err(response) => {
+            error!(?response, "Failed to upgrade User canister wasm");
+            return response;
+        }
     };
 
     let version = wasm.version;